analytics_bot/pyproject.toml
[project]
name = "placebo-analytics"
version = "0.1.0"
description = "Telegram analytics bot — read-only health data analysis"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "matplotlib>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
